requirements.txt
aiohttp==3.7.4
PyJWT==2.1.0
pytest==6.2.4
pytest-asyncio==0.14.0
python-dotenv==1.0.0